refactor: extract durable roundsManager
